Refactored directory structure. Upgraded npm modules.\
