+++ App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HomeScreen from "./src/HomeScreen";
+import HomeScreen from "./src/screens/HomeScreen";
 
 const App = () => {
   return (
