--- __tests__/src/HomeScreen.test.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import {shallow} from 'enzyme';
-import HomeScreen from "../../src/HomeScreen";
-
-describe("HomeScreen", () => {
-    describe("render", () => {
-        it("should render the homescreen", () => {
-            const homeScreen = shallow(<HomeScreen />);
-
-        })
-    })
-})
