+++ package.json
@@ -11,33 +11,35 @@
   },
   "dependencies": {
     "jest-serializer-enzyme": "^1.0.0",
-    "react": "16.11.0",
+    "react": "16.13.1",
     "react-addons-test-utils": "^15.6.2",
     "react-dom": "^16.13.1",
     "react-native": "0.62.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/runtime": "^7.6.2",
-    "@react-native-community/eslint-config": "^1.0.0",
-    "@types/jest": "^24.0.24",
-    "@types/react-native": "^0.62.0",
+    "@babel/core": "^7.10.2",
+    "@babel/runtime": "^7.10.2",
+    "@react-native-community/eslint-config": "^1.1.0",
+    "@types/jest": "^25.2.3",
+    "@types/react-native": "^0.62.13",
     "@types/react-test-renderer": "16.9.2",
-    "@typescript-eslint/eslint-plugin": "^2.27.0",
-    "@typescript-eslint/parser": "^2.27.0",
-    "babel-jest": "^24.9.0",
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
+    "babel-jest": "^26.0.1",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^6.5.1",
-    "jest": "^24.9.0",
+    "eslint": "^7.2.0",
+    "jest": "^26.0.1",
     "jest-enzyme": "^7.1.2",
-    "metro-react-native-babel-preset": "^0.58.0",
-    "prettier": "^2.0.4",
-    "react-test-renderer": "16.11.0",
-    "typescript": "^3.8.3"
+    "metro-react-native-babel-preset": "^0.59.0",
+    "prettier": "^2.0.5",
+    "react-test-renderer": "16.13.1",
+    "typescript": "^3.9.5"
   },
   "jest": {
-    "setupFilesAfterEnv": ["./jest.config.js"],
+    "setupFilesAfterEnv": [
+      "./jest.config.js"
+    ],
     "preset": "react-native",
     "snapshotSerializers": [
       "./node_modules/jest-serializer-enzyme"
