--- src/HomeScreen.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { View, Text } from "react-native";
-
-const HomeScreen = (): JSX.Element => <View>
-    <Text>HomeScreen</Text>
-</View>
-
-export default HomeScreen;
