+++ src/screens/HomeScreen.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+const HomeScreen = (): JSX.Element => <View style={{
+    flex: 1,
+    backgroundColor: "powderblue"
+}}>
+    <Text>HomeScreen</Text>
+</View>
+
+export default HomeScreen;
